Added an automated build process for item models.
Created an automated build process for adding item models to the resource pack. The item models are stored in the "models" property in the JSON file "models.json" in the folder "resourcepack". This might seem silly for something so simple but it'll make adding new models a lot easier (puts everything in one file, which is nice). On the other hand, if I need new functionality for some reason, I'm kind of screwed.